[ARC-595] Move github configuration to utilities
branch: feature/ARC-595-mpyl-basic-cli
